test: add integration test
