+++ .github/workflows/ci.yml
@@ -17,6 +17,21 @@ jobs:
       - name: Unit tests
         run: npm run test:unit
         
+  integration-test:
+    runs-on: ubuntu-latest
+    name: Integration Tests
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v4
+      - name: Setup Node.js
+        uses: actions/setup-node@v3
+        with:
+          node-version: "22.x"
+      - name: Install dependencies
+        run: npm ci
+      - name: Run tests
+        run: npm run test:integration
+
   e2e-test:
     runs-on: ${{ matrix.os }}-latest
     strategy:
@@ -46,6 +61,7 @@ jobs:
     runs-on: ubuntu-latest
     needs: 
       - unit-test
+      - integration-test
       - e2e-test
     if: github.ref == 'refs/heads/master'
     steps:
